Backend/requirements.txt
beautifulsoup4
huggingface-hub
numpy
pypdf
python-dotenv
redis
requests
scikit-learn
sentence-transformers
